+++ release_notes.md
@@ -20,6 +20,8 @@ This section lists features in master, available by [AppVeyor](https://ci.appvey
   - RMQ now supports batch sizes of up to 10 for consuming messages
   - SNS+SQS now supports batch sizes of up to 10 for consuming messages
   - Added support for the Outbox pattern via DepositPost and ClearPostBox
+  - Fixed https://github.com/BrighterCommand/Brighter/issues/156 to allow different exchange types to be set (was broken by support of delayed exchange)
+  
    
 
 ## Release 7.4.0 ##